add parametric option to scatter + fix up plot parametric
An unfortunate thing here is that this allows functions to return either an array with shape (N, 2) or to return (x, y) which is pretty incongruous. But I think being able to return x,y is very reasonable and matplotlib treats the columns as lines. So for the lotka-volterra example for the same function to work for both the phase plot and the timeseries things have to be this way