allow scalar arguments to be kwarg like.
This allows the natural feeling:
```
iplt.imshow(im, vmin = (0, 125), vmax=(125,255)
```
which will create a slider for vmin and vmax

This also allows for more easily chaining multiple commands and passing an indexed controls object as a scalar. Which then allows for this example that was motivating for this change:

```
im = np.random.randn(100*100).reshape(100,100)
fig, axs = plt.subplots(1,2)
axs[1].hist(im.flatten(), alpha=.4)

ctrls  = iplt.imshow(im, vmin=(-1,0), vmax=(0,1), ax = axs[0])
_ = iplt.axvline(ctrls['vmin'], ax = axs[1])
_ = iplt.axvline(ctrls['vmax'], ax = axs[1])
```
--------
Also only append to a figure to the list of what a controller controls if it is not already present. this prevents potential multiple draws draws

prep scalars for scatter - also allow size to be a single number for scatter

plus cleanup docstring a little bit