feature: Added a second image_segmenter class and interactive functions to display segmenter data
Added a image_segmenter_overlayed class which reads a secondary image thats overlayed over the primary. This is useful for medical data for example PET where one has an anatomical reference image and then a heatmap overlayed and wants to segment based on both.
Also added some functions to go along with it as well as a example notebook.
Also had some issues with commiting, these are describes in an issue here: https://github.com/ianhi/mpl-interactions/issues/265.
This changed the pre-commit-config.